pyproject.toml
[tool.poetry]
name = "raresight"
version = "0.1.0"
description = "Early detection of rare dermatological conditions using self-supervised deep learning"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "raresight", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"

# Deep Learning core
torch = {version = "^2.3.0", source = "pytorch"}
torchvision = {version = "^0.18.0", source = "pytorch"}
timm = "^1.0.3"                        # ViT / pretrained models
einops = "^0.8.0"                      # tensor rearranging (MAE)

# Data & ML utilities
numpy = "^1.26.4"
pandas = "^2.2.2"
scikit-learn = "^1.5.0"
Pillow = "^10.3.0"
albumentations = "^1.4.8"             # medical image augmentation
opencv-python-headless = "^4.9.0"

# Experiment tracking
wandb = "^0.17.0"
tensorboard = "^2.17.0"

# Config management
hydra-core = "^1.3.2"
omegaconf = "^2.3.0"

# API
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-multipart = "^0.0.9"
pydantic = "^2.7.3"

# Frontend
streamlit = "^1.35.0"
plotly = "^5.22.0"
altair = "^5.3.0"

# Visualization & analysis
matplotlib = "^3.9.0"
seaborn = "^0.13.2"
grad-cam = "^1.5.2"                   # Grad-CAM explanations

# Utilities
tqdm = "^4.66.4"
rich = "^13.7.1"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-cov = "^5.0.0"
ruff = "^0.4.8"
mypy = "^1.10.0"
ipykernel = "^6.29.4"
notebook = "^7.2.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
select = ["E", "W", "F", "I", "N", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"
